feat: merging agents together
