+++ graph/src/agent/langgraph-studio.py
@@ -3,15 +3,15 @@
 This agent returns a predefined response without using an actual LLM.
 """
 
-from pathlib import Path
+from graph.src.agent.graph import graph
 
-from graph.src.product_research_agent.graph import graph
+
 
 compiled_graph = graph()
 
 if __name__ == "__main__":
     print("Starting the agent...")
-    
+
     while True:
         # Get user input from console
         url_input = input("docs URL: ")
