--- graph/src/product_research_agent/__init__.py
@@ -1,8 +1,0 @@
-"""New LangGraph Agent.
-
-This module defines a custom graph.
-"""
-
-from dotenv import load_dotenv
-
-load_dotenv()
